requirements.txt
runpod
requests
gspread
google-auth
